Move SqliteCacheStorage into scrapy.extensions.httpcache
